move schemas.py to db/schemas.py as it makes more sense, refactor dbapi.py a bit so it now more encapsulated
