+++ src/db/database.py
@@ -5,6 +5,6 @@ from src.config import DATABASE_URL
 
 engine = create_engine(DATABASE_URL)
 
-Session = scoped_session(sessionmaker(bind=engine))
+SessionLocal = scoped_session(sessionmaker(bind=engine))
 
 Base = declarative_base()
