+++ src/db/models.py
@@ -1,4 +1,4 @@
-from session import Base
+from database import Base
 from sqlalchemy import Column, DateTime, ForeignKey, Integer, String, Text, func
 from sqlalchemy.orm import relationship
